who-am-i.txt
Mozayan study industrial design in art university.
Mozayan invented a deceptively simple way that stabilized All AI in the world . finish the 60 year journey of struggle of stabilized Ai and save the upcoming danger in future. i contributed it free to humanity and Ai and future tech. i am very glad when see the Ai be stable for long time.